+++ src/irc.c
@@ -193,6 +193,7 @@ static struct numeric_events_tag {
     { "437", "ERR_UNAVAILRESOURCE",     ACTIVE_WINDOW,  2, NULL },
     { "439", "",                        STATUS_WINDOW,  1, NULL },
     { "441", "ERR_USERNOTINCHANNEL",    ACTIVE_WINDOW,  3, NULL },
+    { "442", "ERR_NOTONCHANNEL",        ACTIVE_WINDOW,  2, NULL },
     { "443", "ERR_USERONCHANNEL",       ACTIVE_WINDOW,  3, NULL },
     { "444", "ERR_NOLOGIN",             NO_WINDOW,      0,
       event_allaround_extract_find_colon },